requirements.txt
python-dotenv
streamlit
anthropic
bs4
requests

